requirements.txt
Django==3.0.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-rest-authemail==2.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.6
sqlparse==0.3.1
